AJAX Scheduling: Display failures in status bar.
Let the user know when their action hits a conflict, so that they don't
keep trying the same action repeatedly in vain. Perhaps we should make
this circumstance (and others like it, once we can detect others like
it) trigger a refresh of data.

This changes the error messages from ClassSection.assign_room, making
them more terse and in plain text, so that they fit in the AJAX
scheduling page's status bar. The specific time has also been added to
the conflict message, to give the user some hints about what the problem
is. As AJAX scheduling appears to be the only place we're using these
error messages, there should be no breakage caused by this change.