Scheduling: ClassSection.assign_room test.
This includes a regression test for the previous commit. If you try to
schedule a class (2) into conflict with an already-scheduled class (1),
then (1)'s schedule should be unaffected and (2)'s should be empty.

I put this in as an AJAXSchedulingModule test, despite the fact that it
only actually tests the API we've built into our models.