Fixed login box appearance on failed login...
If you typed in an incorrect username or password and used the Ajax
login box, you would see a "Hello, !" message in the box after
hitting enter/Go.  This is because the user object gets cast to an
ESPUser... but if the user is None, then a new ESPUser gets created
and the system assumes it's authenticated.  Work around this by
checking that the user has an ID as well in the template.