AJAX Sched.: Backend teacher availability check.
In case there's complicated coteacher structure in the classes and
multiple users are scheduling, verify in the backend that all the
teachers are available when requested before assigning times.

A more disruptive change (to ClassSection.viable_times) would allow a
more informative error message, but in theory we could do that instead
on the client side after refreshing the data.

This also clears a class's meeting time if the room assignment fails.
Otherwise a class might fail to schedule and still cause teachers to be
marked as unavailable.

REMARK TO FUTURE DEVS: If we ever start asking for 'assignreg's without
preceding them with 'deletereg's, then we should replace the invocation
of cls.clear_meeting_times() with something smarter (i.e. something to
restore the previously assigned meeting times).