AJAX Scheduling: Test backend availability check.
This adds a regression test for the preceding commit. It's kind of
loose, in that it doesn't go all the way and check that rooms are
assigned correctly, but that can be added. For now, it at least detects
the change in the previous commit.

Since it makes use of force_availability, I've included a test for that
too. For total correctness it should go in a SchedulingModule test, but
I have doubts that the old SchedulingModule will stick around much
longer.

Some codepaths common to multiple AJAX Scheduling module tests have
moved out to their own methods.

The model-API test now uses classes taught by different teachers, in
case the code for checking availability constraints gets moved around.