Implemented a javascript based hide/view system for inline QSD
Previously, inline QSD editing had checked whether or not the user
was an admin to determine whether or not to show the edit header.
This was a problem because users which were not admins but had QSD
editing bits would not see this even though they were able to edit
QSD.

I fixed this by adding another cookie, cur_qsd_bits, and two new
css classes, qsd_bits and qsd_bits_hidden, which are analogous to
cur_admin, admin, and admin_hidden respectively. The editing
header now appears correctly for users with QSD bits but not admin
privileges.