Add functionality for hiding registration types
Adds a controller for setting (via a Tag) which RegistrationTypes should
appear when a student views his/her inline schedule on the main studentreg
page.

Also adds an auxiliary call for AdminCore for setting and viewing which
RegistrationTypes are wanted. Adds an appropriate Form and template for
making this choice.

The Controller assumes that Enrolled is always a desired RegistrationType,
and forces it to always be set, no matter what the user selects. This is
to prevent it from accidentally being de-selected and causing chaos.

There is also an option to display all RegistrationTypes, which was the
old default. This is so that chapters may choose to keep the old
functionality, if so desired.