Add functionality for hiding registration types
Add the logic for taking the data from the Visible Registration Type
Controller and only displaying the appropriate RegistrationTypes in the
inline schedule on the main studentreg page.

This involved adding a verbs argument to some methods. To not interfere
with pre-existing workflows, the default arguments are such that there
should be no changes to those workflows. Only setting the verbs argument
will yield a behavior different from the default (the default is to list
all valid ClassSections).