Adds support for International Phone Number Format
Resolves github Issue #169.

Typing 1-XXX-XXX-XXXX is now valid, and accepted by our ContactForms.
Previously, the code used a custom PhoneNumberField, which used a regex
that did not support such strings. But it turns out that, not only does
Django have its own USPhoneNumberField, but the regex it uses accepts
1-XXX-XXX-XXXX and XXX-XXX-XXXX as valid phone numbers. So the code now
uses Django's version.

There are other forms in the site which use the custom version. We
should make it a point to phase out the custom version, and integrate
Django's version.