Adds Tag to enable dates in the teacher schedule
Resolves github Issue #86.

Occasionally, there might be a program that occurs on multiple weeks,
but with different classes each week (for example, MIT ESP's ESPrinkler
sub-program). In this case, the default date (something of the form
"Sun, 10:05am -- 10:55am") format doesn't give enough information to
teachers, since they don't know which Sunday they've been scheduled
for. Setting the new friendly_times_with_date Tag to True for a
program will enable the date to be displayed as something of the form
"Sun, July 10, 10:05am -- 10:55am". By default, the Tag is not set,
since most programs are one-weekend, and one-weekend programs won't
want this behavior.