Update TemplateOverrideTest to use reversion.
The New World Order for template overrides is to use the context
manager provided by reversion.create_revision() when updating, and
to use a .revert() method on an appropriate version object when
reverting changes. (Deletion is reserved for wiping out an object
completely.)

This almost fixes tests. Reversion 1.5.1 is still failing two of its
own tests on my dev server.