Lots of test fix-ups.
Basically, because ClassSection is now more effectively denormalized, its calculated fields can't safely be accessed after the database has been modified.  So, in tests, force the relevant Section to be re-queried before checking its value after an update.

I'm not sure this is the API we really want?  On the other hand, trying to have updates be automatic is quite annoying, since there are so many ways to update the database and we can't readily intersect them all...