Add VARNISH_* fields to django_settings.py
I'm not sure we actually want to do this...  Varnish isn't in the default dev-server install, and I feel like it's a somewhat-heavy requirement to add, in that it's always running and poses whatever security issue for the system that's running it if Varnish ever has a security hole and the person running the dev server isn't on top of updating their packages.

Perhaps we want to use unittest2's conditional-test-skip logic to skip the Varnish tests if these variables aren't set, or are set to False or something?