Disable Ajax login from default templates...
The previous few commits updated our default templates to not use
any user-specific information (everything is shown/hidden/rendered
after the page loads, using Javascript).  We had an "Ajax login"
capability that lets you log in without reloading the current page.
This gets more complicated when things are dynamically rendered.
I'm working on code to support it in a separate branch that should
not be included in the upcoming release.

To keep things simple and play nicely with proxy caching,
I disabled the Ajax login scheme for now.  In most cases this shouldn't
be a problem since the page that gets reloaded is simple/easy,
but I'd be interested in your recommendations.