Replace HTTP 302 redirects on login with "meta" redirects...
This is a huge hack and should be considered carefully.

A lot of our pages are proxy-cacheable now thanks to Adam's
work in #206, which is great.  However, we now have a problem
with logins.  When you log in, you might be redirected to a
page (typically /learn/index.html or /teach/index.html) which
is proxy-cacheable.  Thus, when you receive a copy of this page
after being redirected, it does not include the Set-Cookie
headers necessary to tell your browser (and hence our Javascript
code) that you have logged in.  I adjusted our middleware to
allow setting cookies in a response that weren't present in the
request, but this doesn't work by itself because Set-Cookie headers
in an HTTP 302 (redirect) response get ignored.

So, I've replaced our HTTP 302 redirects with "meta" redirects, i.e.
a standard HTTP 200 response with almost no text and a <meta> tag
telling the browser to fetch the next page (e.g. /learn/index.html).
This response includes the cookies and now the browser accepts them,
so that when the following page is loaded, the Javascript can set
the appropriate variables based on your logged-in user.

I would like to find a better workaround and revert most of all of this
change, but this is what came to mind first.