Improve user checking in auth middleware...
The auth middleware was previously programmed to delete
all user-specific cookies if a user was not cached along with
the request or the response.  This is no longer the desired
behavior -- we want the user's cookies to remain as-is so that
they can browse the site (including fully cached pages) while
logged in, and delete them only when the user actually logs out.

Now, in order to log a user out, you have to pass an AnonymousUser
instance via request._cached_user or response._new_user.  This is
done by the signout view added in the previous commit (deecb84).