Update render_to_response() to include request in context...
Several of our templates expect a 'request' variable in the context,
usually for identifying the current user or the path of the request.
We no longer include the user information since we don't want to
trigger any unwanted cookies, but we sometimes need other information
from the request.  One example is the MIT graphic design, which
relies on the request path in order to select the current "tab color"
for different areas of the site.