Change default cache_inclusion_tag context class
The AutoRequestContext causes session data to be accessed, which
leads to Django's session middleware adding a Vary: Cookie header.
We don't want this on pages that need to be proxy cacheable.  So, by
default let's use the standard Context class.

Some cache_inclusion_tag declarations may need to be updated if the
templates that they render need the {{ request }} variable.  The
necessary update is to add the context_class=AutoRequestContext
keyword argument to @cache_inclusion_tag().