Convert MIDDLEWARE_CLASSES to a tuple
This only affects tests.  Some of Django's test code expects the
MIDDLEWARE_CLASSES to be immutable since it modifies the setting
in order to run its tests and then restores the original value.
This does not work if MIDDLEWARE_CLASSES is a list.  Converting to
a tuple eliminated the spurious use of RemoteUserMiddleware that
was breaking our NoVaryOnCookieTest.