Catch empty argument lists in cache_inclusion_tag invalidation
The cache_inclusion_tag relies on the caching API depend_on_cache()
function to cascade invalidation of the underlying cache (i.e. cache
for the function that generates the context dictionary for template
rendering) to the cache that stores HTML rendering results.  When the
entire underlying cache is dumped, we need to notice that and also
dump the entire result cache.  We can detect this by checking for the
empty key set.