Add basic test of ProgramPrintables
Several of the printables views require selecting a user list, so
this test comes with a few pre-selected views (e.g. teacherlist,
student_tickets) and simple lists of students/teachers to try them
with.  There is a function to step through the list of views and
ensure that the printable is generated without error; there are a
couple of specialized functions that add some cursory correctness
checking.  There's a lot of room for improvement but hopefully this
will reduce the surprise of code changes breaking the important
printables without us noticing.