Add script to check all of our Javascript using Closure
Google's Closure compiler can be used to optimize/minify Javascript
code.  We're not using it for that, but we are using it to check if
there are any errors in our code.  This commit adds a script that
attempts to compile all of our non-library Javascript code and see
which files are causing issues.

In order to use this, you need to define CLOSURE_COMPILER_PATH in
local_settings.py and install the binary Closure compiler (compiler.jar)
in that location.