Makes a couple changes to the qsd seltest
I didn't realize that setUp and cleanUp were called before and
after each test. Making use of this I moved the setting and
unsetting of the testserver_port attribute to these, instead of
inside the function that actually checked the page.

Also found a random log call that I missed earlier and deleted it.