Modifies CSRF checking and some tests based on loginbox changes
The new loginbox system broke the old CSRF checking system. It
allowed me to implement a bit of a cleaner and more elegant
solution, though. The system is now to set the csrftoken cookie
on the submission of the form, and then to set the middlewaretoken
afterwards. The check now actually succeeds on the first try
instead of the second in some cases.