Replace {{ MEDIA_URL }} with /media ...
We should probably use {{ MEDIA_URL }} universally, but at the moment
we were only using it in a few places and it's not always present in
the context at template render time.  (That happens because we don't
use RequestContext on views that need to be proxy cached, due to our
paranoia about introducing Vary: Cookie headers.)  I think the best
short-term solution is to hardcode /media like we were already doing
90% of the time.  Longer term, I bet we'll work out a way to use
RequestContext and switch everything to {{ MEDIA_URL }} so we have
flexibility over who serves static files.