Fix login redirect bug
We were catching Django's default 302 redirects on login in order to
rewrite their location and replace them with a 200 META redirect
(necessary for setting cookies).  However, we need to do this whenever
Django issues a 302 redirect, even if we're happy with the URL
that this redirect is redirecting to.

This change allows you to log in from a QSD page that has been loaded
from cache, and be returned to the same page (again loaded from cache)
with it showing that you've logged in.