Tweak profile grade display behavior.
We added code to display the user's current grade for informational
purposes if it is out of range.  This was displaying '0' for new
accounts, so I added a check to only show the current grade if the
user has at least 1 registration profile already.

Also, it turned out this didn't happen when the student_grade_options
Tag was set.  I rearranged the code so that the current grade is added
to the list of options whether or not the Tag is set.