Allow overriding catalog sort order with a keyword argument
The ClassManager.catalog_cached() function is now used for the
dashboard as well as the class catalog.  The sort order can be
controlled with the catalog_sort_fields Tag, but we want it to
be different for the dashboard.  So I added an argument to
catalog_cached() which overrides both the default and the Tag,
and used this to sort the classes in the dashboard by ID.