Update our custom e-mail backend to Django 1.3
We had been overriding SMTPConnection in order to set a return path
distinct from the from address on our e-mails.  This commit updates
to subclass their newer backend API so that we can easily switch
between our backend and the default backends.  This is required for
testing our e-mail systems (using Django's in-memory e-mail backend).