Fix bug in teacherreg grade range popup.
When creating an open class, grade range options are not given to the
user, and the grade range is set to the min and max grade of the
program. This was causing the grade range pop-up to activate, even
though there is no way for a teacher of an open class to change their
grade range to something smaller. So now the pop-up will not be
activated when creating or editing an open class.