Disable auto-setting of teacher availability
The teacher availability form defaults to "100% availability"
when the current specified availability is empty, or not enough
to cover the registered classes.  However, there was code to
actually save this 100% availability in addition to populating
the form.  So, if you left the page without clicking "submit"
on the form, your availability in the database might not match
what you expected!  I commented out the relevant lines of code...
maybe this deserves to be wrapped in a Tag.