Beef up program module decorators with login checking
We have decorators that check properties of the current user in
order to determine whether to show the page or an error message
(e.g. "you are not a teacher", "registration is not yet open").
Some of these decorators check whether the user is logged in
and return a login page (which will forward to the desired view
after a successful login) if they are not yet logged in.  I
modified the check to avoid our AnonymousUser issue (apparently
user.is_authenticated() returns True) by checking the user's ID,
and extended the check to the @meets_deadline family of decorators
as well as the @needs_* ones.

I'm not sure why this was necessary, any ideas?