Fix bug in module checkbox template
Previously, the template was checking to see the value of
module.module.required (i.e., the default value of required for the
ProgramModule) rather than module.required (i.e., the value of required
for the ProgramModuleObj for that particular program). This meant that
making a module required or not required wouldn't affect the text label
indicating whether it was required or not required. This commit changes
that, so now module.required is correctly referenced.