Fixes logged out bug in program modules
Trying to go to a program module's page causes
ProgramModuleObj.findModule() to be called. If the module in question
is a core module, the code loops through all required modules of the
same type, and calls module.isComplete() on each of them. This was
happening even if the user was logged out. But module.isComplete() will
fail in almost all cases when the user is logged out, because an
AnonymousUser does not have any data associated with it.
This was causing requests for /teacherreg/, /studentreg/, etc. to fail
when accessed by a logged out user, rather than gracefully redirecting
to the login page. This commit fixes that behavior, by checking if the
user is logged in if the module is a core module.

The access any non-core module while logged out should already be
handled correctly by the @needs_* method decorators. But this is not,
and should not, be the case for ProgramModuleObj.findModule() or
module.isComplete().