Check for overcommitted teachers in coteacher interface
A teacher can get "stuck" if they are signed up to teach more hours of
class than the program offers, so we try to avoid this situation.  We
already had checks in the makeaclass code for this.  However, it was
still possible to overcommit a teacher by adding them as a coteacher to
a class that pushed them over the limit.

This change closes the loophole by noticing this condition as a schedule
conflict and preventing you from adding the coteacher.