Add cache dependency to render_inline_qsd
In most cases the input_anchor argument to render_inline_qsd() is a
DataTree instance.  However, it can also be a string in some places
where we don't want to bother fetching a node and adding it to the
context.  When a QSD is edited we need to flush the caches for inline
QSD tags that were originally referenced in either fashion.