Makes TeacherClassRegModule.makeaclass a main_call
This used to be true, but TeacherClassRegModule.makeaclass() did not
have the @main_call method decorator. So when 'main_call' was removed
from the program module properties in commit
5c1a4f79898d5f71c59af14e5d9aa16ce440f9b6,
TeacherClassRegModule.makeaclass() lost its status as a main_call.

This commit replaces the @aux_call method decorator with the @main_call
method decorator, which makes TeacherClassRegModule.makeaclass() a main
call again. This fixes the bug that was causing
TeacherClassRegModule.makeaclass() not to behave as a required module
view.