Don't fail out if we can't add the customforms schema; just print a warning.
SQLite doesn't support multiple schemas, as far as I can tell.  This of course breaks customforms.  But it doesn't break anything else.  And SQLite is useful.  So don't fail outright if we can't do multiple schemas.

It'd be nice if we could get customforms to work with SQLite someday.  Multiple databases are cheap and easy these days; perhaps, with the SQLite backend, it could just open a second database?