Don't create the same index twice, in this migration
For some reason, with SQLite, the South alter_table() statement seems to try to create indices once for each time a field with an index and with db_index=True is altered.  (ForeignKeys default to db_index=True.)  We're playing around with this column a lot in this migration, but we only want to create the index once; so for all intermediate alter_table()'s, force db_index=False.  (Creating the index at the intermediate points is strictly useless because South defers index creation; it creates indices all at the end, so it was literally executing three consecutive identical CREATE INDEX statements.)