Generic workarounds mechanism for adding support for missing PL/PgSQL procedures
I've added a global dictionary PROC_BACKUPS; it maps (string) procedure names to Python procedures that return QuerySets of the same type as the specified procedure is supposed to return.

I've also added in a backup for userbit__bits_get_user.  Partly as an example, partly since that seems to be by far the most common function for us to call.