Work around bug in dashboard: number of sections per category
The query that fetches the number of classes and sections per category
was returning incorrect number of sections for "walk-in seminars" when
"walk-in seminars" was a listed category of the program.  I had
difficulty adjusting the query to work properly, so to work around the
issue I introduced another query to find the distinct list of valid
categories.  The main query finds the totals for every category, and
then the invalid categories are filtered out.

Better solutions are welcome.