Adds StrippedCharField
The StrippedCharField strips whitespace from the submitted value before
it is validated. Thus, submitting a value with only whitespace causes
the form submittion to fail if the field is required. It is also a
subclass of SizedCharField, that way you can easily size the widget,
and instances of SizedCharField can be easily changed to instances of
StrippedCharField. This commit makes some of these changes, in order to
prevent users from submitting blank values in required fields of the
user creation, user profile, and class creation forms.

This commit reverts the validation that was added in ba1d23727bf8539be,
since those functions are no longer needed. The fields are instead
replaced with instances of StrippedCharField