Work around error in Ajax scheduling on data refresh
I don't fully understand why this is happening, but sometimes when the
RoomFilter is saved/restored, some of its member variables (including
minsize and maxsize) get clobbered.  Then when the filter() function is
called and attempts to check those variables, we get an error (Matrix.js
line 323).

This change checks for existence of the necessary variables in order to
avoid hitting Javascript errors (leading to a blank page after data is
refreshed from the server).  However, it does not fix the underlying
problem.  Further attention is needed.