Manage Class: reject sections if rejecting class.
In line 337, there was a comment "If the parent class was canceled,
cancel the sections"---followed by code that set section statuses
to what the parent class's status used to be.

(It actually worked for a while; the problem arose from a surprise
interaction between 614a7794 (saving class info after section info) and
ec9ba3b2 == 9b9b3f2e (rejecting sections when rejecting the class).)

This probably deserves a regression test, but AdminClass's forms are
enormous so I'll just say for now that I tested it in my browser and
will add a test to the dev branch later.