Switch from Django send_mail() function to our own
Our own send_mail() function copies to the archive address by default,
which is helpful.  It also includes a Return-Path header to deal with
bounces.  We may as well use it by default.

I left a few instances of the original send_mail() function in, for
instances where there could be confidential information in the message
(in which case it shouldn't be archived)... and I think one place that
there was a circular import when using our own function.