Add ClassSubject cache dependency to ClassSubject.teachers()
This should fix the problem of newly created classes not having teachers,
which may have been caused by the anchor being changed after the cache
was populated (causing the selector function for the existing UserBit
dependency to return the 'wrong' thing).

Also, while looking into this I found tons of old caching stuff in
the Class{Subject,Section} models that needs to be eliminated, but not
right now.