Adds redirects to manage pages in admin classedit
When you are an admin editing a teacher's class, you can now choose to
be redirected either to the same page, the dashboard, or the admin
portal after saving your changes. This commit also adds links to the
dashboard and the admin portal at the top of the admin editclass page.