Fixes required module bug caused by 2+ main_call
When a required, uncompleted program module is found, the function
ProgramModuleObj.main_view_fn() is called. This function is a wrapper
for that module's main_call method.

Previously, the main_call was determined by self.main_view. When a
module (such as the TeacherEventsModule) has two main_calls,
self.main_view doesn't necessarily return the view you want. The
observed bug was that, when the Teacher Training and Interview Signups
ProgramModuleObj was made required, teachers would be redirected to the
Manage Teacher Training and Interviews view instead.

This commit replaces the reference to self.main_view with a reference to
self.get_main_view(tl), which was introduced in commit b11641b29d and
returns the correct main_call view, even in cases when there are
multiple main_calls.