Convert ClassSection.get_meeting_times() to use caching API
The existing function used local memory storage for the cached result
of self.meeting_times.all(), so in rare cases the result of
get_meeting_times() was non-empty when in reality the real list of
meeting_times was empty.  This caused errors in
ClassSection.initial_rooms().  Fixed by adding @cache_function and
depend_on_m2m.