Fixes customform templates to import jQuery once
Makes use of the new blocks in the elements/html template in order to
override the version of jQuery that is imported and binded to $j.

In the case of customforms/form_base.html and
media/scripts/customforms-validation.js, this change correctly loads
two different versions of jQuery, and binds them to distinct variables.
The usual jQuery version is, as always, bound to $j, whereas the
version that customforms uses is bound to $jcf.