Add lock_level field to ResourceAssignment ...
Motivated by the needs of automatic scheduling, this change allows
you to specify just how 'sticky' you want a schedule assignment to
be.  Each resource assignment now has an integer "lock level" which
is 0 by default; this will usually be interpreted as "OK to change
via automatic scheduling." Higher lock levels can be used to specify
that some assignments should not be changed.