Create TeacherCheckinModule
Teachers are marked as checked-in by giving them the
V/Flags/Registration/Teacher/Arrived userbit on the program they're
checking-in for.  The userbit expires at the end of the day.

missingteachers lists the sections where at least one teacher has not
checked-in, along with all the teachers for the section.