Add caching to JSON views
This uses the @cached_module_view decorator so that we can ensure the
results are correct (and also check that the appropriate type of user is
logged in).  For the class_info view we probably don't care and
can treat the results as static, much like we do for the catalog.