Adds FileFields to customforms
This code includes an arguably important TODO. Since the current
version of the Django FormWizard only allows FileFields on the last
page of a multi-page form, technically this code allows the user to
create a seamingly-valid customform that will actually cause Django to
raise an Exception upon its rendering. For now, since the only custom
forms I've ever seen used are one page, this shouldn't be an issue. But
I want to come back to this in the future, when I have more time, and
write code that will prevent this from happening.