Add CustomFormModule for use as "extra profile information" step
This is a slightly generalized adaptation of the teacher quiz: directors
can create a custom form, provide its ID to the Tag 'learn_extraform_id'
or 'teach_extraform_id' and ask participants to fill it out as part of
registration.  This will allow our chapters to add questions of their
choice to registration without having us modify the code, so long as what
they need can be represented by a custom form.