Prevent parallel processing of MessageRequests
In our cronmail process_messages() function we look for message requests
that have not yet been processed, mark them as processed and then send the
e-mails.  If a second script is started in parallel and checks for message
requests before the first script has marked them all as processed, then
the message request could be processed twice, creating duplicate
TextOfEmail objects.

This commit adds a command that gets translated to an SQL UPDATE statement
which sets the processed_pid field of the appropriate message requests, so
that even cron jobs running in parallel will not process the same message
request more than once.