Send e-mails only for messages just processed
It was possible for two dbmail_cron scripts running in parallel to send
overlapping sets of e-mails.  The TextOfEmail objects would be identified
by one instance, then the other instance would look for un-sent
TextOfEmails before it was done marking them as sent.  There are two
changes here to address this:
- In line with the previous change to atomically update MessageRequests,
  the list of MessageRequests that have been processed by each script
  instance is padded to the send_email_requests() function and only
  TextOfEmails pertaining to those requests are sent.
- The TextOfEmail objects are marked as sent by a bulk update() statement
  which happens faster than iterating over all the objects and calling
  save().