View teachers who did or did not check in on a previous day
Add a form at the bottom of the teachercheckin page to select a date
and time, which causes missingteachers to look for userbits for check-
ins that were active at that time instead of the current time.

Remove the form for checking in teachers from teachercheckin, because
all of its funcionality (and more) is available from the
missingteachers page.  The form autocompleted to every user on the
website, and there's no way to fix it without rewriting the backend
for AjaxForeignKeyNewformField.