Add a new, faster lottery assignment script.
The script does its own handling of schedule constraints and lunch
constraints by keeping track of the schedule and preference information in
numpy arrays.  This is noticeably faster than using the API functions e.g.
ClassSection.cannotAdd().

The script is written in the form of a controller:
  LotteryAssignmentController in esp/esp/program/controllers/lottery.py

To use it:
1. Initialize the controller:
     lac = LotteryAssignmentController(my_program)
2. Compute lottery assignments:
     lac.compute_assignments()
   (Note that the results are checked for consistency each time you do
   this.)
3. Obtain statistics on the results:
     stats = lac.compute_stats()
4. Review the stats and, if desired, return to step 2 as many times as you
   want.
5. Once you are happy with the way things look, save the results (e.g.
   StudentRegistration bits):
     lac.save_assignments()