Update customforms with FormWizard from Django 1.4
The FormWizard from Django 1.3 is unable to accommodate File Fields on
all but the last step of the form. The one in Django 1.4 allows us to
insert File Fields into all steps of the form. This commit plugs in
the new FormWizard as an app and updates the customforms code to play
nicely with it. This should hold until the codebase is updated to Django
1.4.

The app must be installed using: pip install django-formwizard.