Fix bug in fillslot page with lunch constraints
If you were missing a lunch period and went to a fillslot page for a
timeslot that would require you to have a lunch period, the first class
listed on the fillslot page would claim to violate the schedule
constraints and all subsequent classes would be available.  It turned out
that the fillslot page was checking the constraints for each class (good)
but also auto-adding a lunch period if the constraints weren't satisfied
(bad).  This change puts an end to that behavior.