Create UserSearchController
We have some messy query code in esp/esp/users/views/usersearch.py.  This
commit takes the query logic of the search_for_user function (which is
used to filter down a list of users based on criteria like IDs, zip codes
and graduation years) and puts it in a controller so that other parts of
the code can use the functionality, without necessarily following the
standard view sequence used for generating a user list.