Adjust algorithm for finding students for onsite class list
This now uses the same set of students as the "All students who have
filled out a profile" list.  It's now possible to use combinations with
other lists (e.g. class/lottery registration), but I found that the
database time needed for that was prohibitive for large programs.  We may
just need to require that students fill out a profile.

References issue #115.