Add user filtering options to new comm panel
This includes filtering by name, school, etc.  Also tries to show only the
options you need, i.e. no need to enter in a school if you're searching
for parents.  In the future this might let us surreptitiously show
different fields which are called the same thing, but look in different
parts of the database, e.g. "school" for students vs. "school" for
educators.